Removed GPS readout from moving map view. Rounded the heading and padded it with leading zeros. Changed capitalization of ground speed units. Removed DAFIF. Repaired data-type problem in terrain profiler. Updated to version 1.004.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@2629 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWindIOS/TAIGA/TAIGA/TerrainProfileView.m
@@ -32,7 +32,7 @@
     UILabel* rightLabelView;
     UILabel* noCourseLabel;
     UILabel* aircraftAltitudeLabelView;
-    float* gradientColors;
+    CGFloat* gradientColors;
 }
 
 - (TerrainProfileView*) initWithFrame:(CGRect)frame worldWindView:(WorldWindView*)worldWindView
@@ -46,7 +46,7 @@
     ys = nil;
 
     _opacity = 0.5;
-    gradientColors = malloc((size_t) 24 * sizeof(float));
+    gradientColors = malloc((size_t) 24 * sizeof(CGFloat));
     gradientColors[0] = 0.0;
     gradientColors[1] = 1.0;
     gradientColors[2] = 0.0;
